フィードバック内容をもとに修正
