+++ src/components/profileImage/ProfileImage.module.scss
@@ -1,9 +1,10 @@
 .root {
   width: 100%;
-  height: 200px;
+  height: 300px;
   .coverPhotoWrapper {
     background: url(./Rectangle81.jpg);
     background-repeat: no-repeat;
+    background-size: contain;
     width: 100%;
     height: 100%;
   }
