+++ src/components/profileMainImage/ProfileMainImage.module.scss
@@ -2,7 +2,7 @@
   height: 120px;
   width: 120px;
   .mainPhotoWrapper {
-    background: url(./L.jpg);
+    background: url(./L.png);
     background-repeat: no-repeat;
     width: 100%;
     height: 100%;
