+++ src/data/Account.ts
@@ -1,4 +1,4 @@
-import { Profile } from "./Profile";
+import { Profile } from "./profile/index";
 
 export interface Account {
   id: string;
