--- src/data/Profile.ts
@@ -1,5 +1,0 @@
-export interface Profile {
-  id: string;
-  firstName: string;
-  lastName: string;
-}
