+++ src/data/academicHistory/index.ts
@@ -1,0 +1,8 @@
+export type AcademicHistories = {
+  id: string;
+  name: string;
+  faculty: string;
+  sinceDate: string;
+  untilDate: string;
+  type: string;
+};
