+++ src/data/workHistory/index.ts
@@ -1,0 +1,19 @@
+export type WorkHistories = {
+  id: string;
+  isEmployed: string;
+  occupation: {
+    id: string;
+    name: string;
+  };
+  industry: {
+    id: string;
+    name: string;
+  };
+  position: string;
+  annualIncome: number;
+  managementExperience: number;
+  jobSummary: string;
+  sinceDate: string;
+  untilDate: string;
+  name: string;
+};
